numstore_updated/Numstore-main/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
asyncpg>=0.29.0
bcrypt>=4.1.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
jinja2>=3.1.4
starlette>=0.37.0
itsdangerous>=2.1.2
resend>=2.0.0
pydantic[email]>=2.0.0
httpx>=0.27.0